Added date attributes. These are specified in the arff header as:
@attribute <name> date [<format>]

<name> is the name for the attribute
<format> is an optional string specifying how date values should be
parsed/printed (this is the same format used by SimpleDateFormat). The
default format string accepts the ISO-8601 combined date and time
format: "yyyy-MM-dd'T'HH:mm:ss".

and specified in the data section as the corresponding string
representations of the date/time. Here's a simple example:

@RELATION timetest
@ATTRIBUTE timestamp DATE "yyyy-MM-dd'T'HH:mm:ss"
@DATA
2001-04-03T12:12:12
2001-05-03T12:59:55

Date attributes return true for isNumeric(). (the double value stored
in each Instance is the double representation of Date.getTime()).

The database support (in InstanceQuery) for dates is implemented but
needs to be tested.


git-svn-id: https://svn.cms.waikato.ac.nz/svn/weka/trunk@1049 e0a1b77d-ad91-4216-81b1-defd5f83fa92